adds react-hook-form, yup resolver and form to add new employees

+++ employee-frontend/src/components/Employee/Employee.tsx
@@ -31,16 +31,30 @@ const Employee: React.FC<EmployeeProps> = ({ employee }) => {
         finishDate,
     } = employee;
 
-    const { refresh, setRefresh } = useContext(RefreshContext);
+    // required to avoid the TS error due to the possible null value
+    const refreshContext = useContext(RefreshContext);
+
+    if (!refreshContext)
+        throw new Error(
+            "RefreshContext must be used within a RefreshContextProvider"
+        );
+    const { refresh, setRefresh } = refreshContext;
 
     const handleDelete = async () => {
-        await deleteEmployees(id);
-        console.log("employee deleted");
-        setRefresh(refresh + 1);
+        const isConfirmed = window.confirm(
+            "Are you sure you want to delete the selected employee?"
+        );
+        if (isConfirmed) {
+            await deleteEmployees(id);
+            console.log("employee deleted");
+            setRefresh(refresh + 1);
+        } else {
+            console.log("Employee deletion cancelled");
+        }
     };
     return (
         <div className="bg-white border border-gray-200 rounded-lg shadow-md p-4">
-            <h3>
+            <h3 className="text-xl font-bold">
                 {firstName} {lastName}
             </h3>
             <p>{id}</p>
@@ -52,7 +66,9 @@ const Employee: React.FC<EmployeeProps> = ({ employee }) => {
             <p>{startDate}</p>
             <p>{finishDate && <span>Finish date: {finishDate}</span>}</p>
             <br />
-            <button onClick={handleDelete}>Delete Employee</button>
+            <div className="mx-auto bg-red-600 rounded-full bg-red w-40">
+                <button onClick={handleDelete}>Delete Employee</button>
+            </div>
         </div>
     );
 };
